chore: remove precise locking of lodash dependency
Use the caret (^) operator to allow any version of lodash with a version over the specified
version (4.17.13).  This means that when this package is installed alongside other packages
depending on a higher version of lodash, the shared lodash dependencies can be deduped properly
across all dependents.

The caret also ensures that this package will not automatically upgrade to 5.x versions of
lodash (the next major version).